requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.7
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
email_validator==2.2.0
python-magic==0.4.27

